+++ background/main.js
@@ -76,6 +76,9 @@ chrome.runtime.onMessageExternal.addListener(
           case 'CityMapService':
             cityMap.process(response.requestMethod, response.responseData, response.requestId)
             break
+          case 'ClanBattleService':
+            clanBattle.process(response.requestMethod, response.responseData, response.requestId)
+            break
           case 'GreatBuildingsService':
             greatBuilding.process(response.requestMethod, response.responseData, response.requestId)
             break